+++ eggs/Marshmallow/src/main/java/com/android_m/egg/preview/PlatLogoActivity.java
@@ -14,19 +14,13 @@
  * limitations under the License.
  */
 
-package com.android_m.egg;
+package com.android_m.egg.preview;
 
-import android.animation.Animator;
-import android.animation.ObjectAnimator;
 import android.app.Activity;
 import android.content.ActivityNotFoundException;
 import android.content.Intent;
 import android.content.res.ColorStateList;
-import android.graphics.Canvas;
-import android.graphics.ColorFilter;
 import android.graphics.Outline;
-import android.graphics.Paint;
-import android.graphics.PixelFormat;
 import android.graphics.drawable.Drawable;
 import android.graphics.drawable.RippleDrawable;
 import android.os.Bundle;
@@ -39,9 +33,7 @@ import android.view.ViewOutlineProvider;
 import android.view.animation.PathInterpolator;
 import android.widget.FrameLayout;
 
-import androidx.annotation.ColorInt;
-import androidx.annotation.Nullable;
-
+import com.android_m.egg.R;
 import com.dede.basic.SpUtils;
 
 public class PlatLogoActivity extends Activity {
@@ -63,7 +55,7 @@ public class PlatLogoActivity extends Activity {
         final DisplayMetrics dm = getResources().getDisplayMetrics();
         final float dp = dm.density;
         final int size = (int)
-                (Math.min(Math.min(dm.widthPixels, dm.heightPixels), 600 * dp) - 100 * dp);
+                (Math.min(Math.min(dm.widthPixels, dm.heightPixels), 600*dp) - 100*dp);
 
         final View im = new View(this);
         im.setTranslationZ(20);
@@ -73,56 +65,19 @@ public class PlatLogoActivity extends Activity {
         im.setOutlineProvider(new ViewOutlineProvider() {
             @Override
             public void getOutline(View view, Outline outline) {
-                final int pad = (int) (8 * dp);
-                outline.setOval(pad, pad, view.getWidth() - pad, view.getHeight() - pad);
+                final int pad = (int) (8*dp);
+                outline.setOval(pad, pad, view.getWidth()-pad, view.getHeight()-pad);
             }
         });
-        final float hue = (float) Math.random();
-        final Paint bgPaint = new Paint(Paint.ANTI_ALIAS_FLAG);
-        bgPaint.setColor(HSBtoColor(hue, 0.4f, 1f));
-        final Paint fgPaint = new Paint(Paint.ANTI_ALIAS_FLAG);
-        fgPaint.setColor(HSBtoColor(hue, 0.5f, 1f));
-        final Drawable M = getDrawable(R.drawable.m_platlogo_m);
-        final Drawable platlogo = new Drawable() {
-            @Override
-            public void setAlpha(int alpha) {
-            }
-
-            @Override
-            public void setColorFilter(@Nullable ColorFilter colorFilter) {
-            }
-
-            @Override
-            public int getOpacity() {
-                return PixelFormat.TRANSLUCENT;
-            }
-
-            @Override
-            public void draw(Canvas c) {
-                final float r = c.getWidth() / 2f;
-                c.drawCircle(r, r, r, bgPaint);
-                c.drawArc(0, 0, 2 * r, 2 * r, 135, 180, false, fgPaint);
-                M.setBounds(0, 0, c.getWidth(), c.getHeight());
-                M.draw(c);
-            }
-        };
+        final Drawable platlogo = getDrawable(R.drawable.m_platlogo_preview);
         im.setBackground(new RippleDrawable(
                 ColorStateList.valueOf(0xFFFFFFFF),
                 platlogo,
                 null));
-        im.setOutlineProvider(new ViewOutlineProvider() {
-            @Override
-            public void getOutline(View view, Outline outline) {
-                outline.setOval(0, 0, view.getWidth(), view.getHeight());
-            }
-        });
         im.setClickable(true);
         im.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                if (mTapCount == 0) {
-                    showMarshmallow(im);
-                }
                 im.setOnLongClickListener(new View.OnLongClickListener() {
                     @Override
                     public boolean onLongClick(View v) {
@@ -143,10 +98,7 @@ public class PlatLogoActivity extends Activity {
                             @Override
                             public void run() {
                                 try {
-                                    startActivity(new Intent(PlatLogoActivity.this, MLandActivity.class));
-//                                            .setFlags(Intent.FLAG_ACTIVITY_NEW_TASK
-//                                                    | Intent.FLAG_ACTIVITY_CLEAR_TASK
-//                                                    | Intent.FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS));
+                                    startActivity(new Intent(PlatLogoActivity.this, ShruggyActivity.class));
                                 } catch (ActivityNotFoundException ex) {
                                     Log.e("PlatLogoActivity", "No more eggs.");
                                 }
@@ -167,9 +119,6 @@ public class PlatLogoActivity extends Activity {
             @Override
             public boolean onKey(View v, int keyCode, KeyEvent event) {
                 if (keyCode != KeyEvent.KEYCODE_BACK && event.getAction() == KeyEvent.ACTION_DOWN) {
-                    if (mKeyCount == 0) {
-                        showMarshmallow(im);
-                    }
                     ++mKeyCount;
                     if (mKeyCount > 2) {
                         if (mTapCount > 5) {
@@ -192,88 +141,5 @@ public class PlatLogoActivity extends Activity {
                 .setDuration(500)
                 .setStartDelay(800)
                 .start();
-    }
-
-    public void showMarshmallow(View im) {
-        final Drawable fg = getDrawable(R.drawable.m_platlogo);
-        fg.setBounds(0, 0, im.getWidth(), im.getHeight());
-        fg.setAlpha(0);
-        im.getOverlay().add(fg);
-
-        final Animator fadeIn = ObjectAnimator.ofInt(fg, "alpha", 255);
-        fadeIn.setInterpolator(mInterpolator);
-        fadeIn.setDuration(300);
-        fadeIn.start();
-    }
-
-
-    public static float constrain(float amount, float low, float high) {
-        return amount < low ? low : (amount > high ? high : amount);
-    }
-
-    /**
-     * Convert HSB components to an ARGB color. Alpha set to 0xFF.
-     * hsv[0] is Hue [0 .. 1)
-     * hsv[1] is Saturation [0...1]
-     * hsv[2] is Value [0...1]
-     * If hsv values are out of range, they are pinned.
-     *
-     * @param h Hue component
-     * @param s Saturation component
-     * @param b Brightness component
-     * @return the resulting argb color
-     */
-    @ColorInt
-    public static int HSBtoColor(float h, float s, float b) {
-        h = constrain(h, 0.0f, 1.0f);
-        s = constrain(s, 0.0f, 1.0f);
-        b = constrain(b, 0.0f, 1.0f);
-
-        float red = 0.0f;
-        float green = 0.0f;
-        float blue = 0.0f;
-
-        final float hf = (h - (int) h) * 6.0f;
-        final int ihf = (int) hf;
-        final float f = hf - ihf;
-        final float pv = b * (1.0f - s);
-        final float qv = b * (1.0f - s * f);
-        final float tv = b * (1.0f - s * (1.0f - f));
-
-        switch (ihf) {
-            case 0:         // Red is the dominant color
-                red = b;
-                green = tv;
-                blue = pv;
-                break;
-            case 1:         // Green is the dominant color
-                red = qv;
-                green = b;
-                blue = pv;
-                break;
-            case 2:
-                red = pv;
-                green = b;
-                blue = tv;
-                break;
-            case 3:         // Blue is the dominant color
-                red = pv;
-                green = qv;
-                blue = b;
-                break;
-            case 4:
-                red = tv;
-                green = pv;
-                blue = b;
-                break;
-            case 5:         // Red is the dominant color
-                red = b;
-                green = pv;
-                blue = qv;
-                break;
-        }
-
-        return 0xFF000000 | (((int) (red * 255.0f)) << 16) |
-                (((int) (green * 255.0f)) << 8) | ((int) (blue * 255.0f));
     }
 }